# packages/api/docs/openAPI.yml
openapi: 3.0.2
info:
  title: Mangas
  version: 0.0.0
paths:
  /api/mirrors/list:
    get:
      tags:
        - mirror
      description: Get information about all mirrors
      operationId: getMirrorsInfo
      responses:
        '201':
          description: 201 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/getMirrorsResponse'
  /api/mirrors/{mirror}:
    get:
      tags:
        - mirror
      description: Get information about a specific mirror
      operationId: getMirrorInfo
      parameters:
        - name: mirror
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getMirrorsResponse'
  /api/mirrors/{mirror}/options:
    get:
      tags:
        - mirror
      description: Get a mirror options list
      operationId: getMirrorOptions
      parameters:
        - name: mirror
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/options'
        '404':
          description: 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
    post:
      tags:
        - mirror
      description: Set options for a mirror
      operationId: SetMirrorOptions
      parameters:
        - name: mirror
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/options'
        required: true
      responses:
        '201':
          description: 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/options'
        '400':
          description: 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400'
        '404':
          description: 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
  /api/mirrors/search:
    post:
      tags:
        - mirror
      description: Search mangas through all enabled mirrors
      operationId: searchManga
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/query'
        required: true
      responses:
        '201':
          description: Results may be an array of search results AND error messages
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/searchMangaResponseError'
                  - type: array
                    items:
                      $ref: '#/components/schemas/searchMangaResponse'
  /api/mirrors/{mirror}/search:
    post:
      tags:
        - mirror
      description: Search mangas in a specific mirror
      operationId: searchMangaInMirror
      parameters:
        - name: mirror
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/query'
        required: true
      responses:
        '201':
          description: 201 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        '400':
          description: 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400'
        '404':
          description: 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
        '500':
          description: 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchErrorMessage'
  /api/mirrors/{mirror}/manga:
    post:
      tags:
        - mirror
      description: >-
        Get a manga information such as title, author, description, chapters
        list, etc.
      operationId: getManga
      parameters:
        - name: mirror
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: relative link to manga page, eg. /manga/one_piece/
              required:
                - query
        required: true
      responses:
        '201':
          description: 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MangaPage'
        '400':
          description: 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400'
        '404':
          description: 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
        '500':
          description: 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MangaErrorMessage'
  /api/mirrors/{mirror}/chapter:
    post:
      tags:
        - mirror
      description: Get chapter's images
      operationId: getChapterImages
      parameters:
        - name: mirror
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: relative link to the chapter, eg. /manga/one_piece/chapter5
              required:
                - query
        required: true
      responses:
        '201':
          description: Response may be an array of chapter pages AND error messages
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/ChapterPage'
                  - type: array
                    items:
                      $ref: '#/components/schemas/ChapterPageErrorMessage'
        '400':
          description: 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400'
        '404':
          description: 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
        '500':
          description: 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterErrorMessage'
components:
  schemas:
    getMirrorsResponse:
      type: object
      properties:
        name:
          type: string
          description: The mirror name
        host:
          type: string
          description: The mirror webpage
        enabled:
          type: boolean
          description: Whether the mirror is enabled or not
        icon:
          type: string
          description: The mirror icon
      required:
        - name
        - host
        - enabled
        - icon
    error404:
      type: object
      properties:
        error:
          type: string
          enum:
            - mirror_not_found
      required:
        - error
    obj:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
    options:
      type: object
      properties:
        option_name:
          oneOf:
            - type: string
            - type: boolean
            - type: number
              format: float
            - $ref: '#/components/schemas/obj'
            - type: array
              items:
                type: string
            - type: array
              items:
                type: boolean
            - type: array
              items:
                type: number
                format: float
            - type: array
              items:
                $ref: '#/components/schemas/obj'
          description: 'key: value'
      required:
        - option_name
    error400:
      type: object
      properties:
        error:
          type: string
          description: error describing why this is a bad request
      required:
        - error
    query:
      type: object
      properties:
        query:
          type: string
      required:
        - query
    SearchErrorMessage:
      type: object
      properties:
        error:
          type: string
          enum:
            - search_error
            - search_error_unknown
        trace:
          type: string
      required:
        - error
    searchMangaResponseError:
      type: object
      properties:
        error:
          type: string
          enum:
            - search_error
            - search_error_unknown
          description: The error message
        trace:
          type: string
          description: The error trace
        mirror:
          type: string
          description: The mirror where the error occured
      required:
        - error
        - mirror
    SearchResult:
      type: object
      properties:
        lang:
          type: string
        name:
          type: string
        link:
          type: string
        cover:
          type: string
        synopsis:
          type: string
        last_release:
          type: object
          properties:
            name:
              type: string
            volume:
              type: number
              format: float
            chapter:
              type: number
              format: float
      required:
        - lang
        - name
        - link
    searchMangaResponse:
      type: object
      properties:
        mirror:
          type: string
          description: Mirror name
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
          description: Array of results
      required:
        - mirror
        - results
    MangaPage:
      type: object
      properties:
        langs:
          type: array
          items:
            type: string
        name:
          type: string
        mirror:
          type: string
        synopsis:
          type: string
        covers:
          type: array
          items:
            type: string
        chapters:
          type: array
          items:
            type: object
            properties:
              number:
                type: number
                format: float
              volume:
                type: number
                format: float
              link:
                type: string
              name:
                type: string
            required:
              - number
              - link
      required:
        - langs
        - name
        - mirror
        - chapters
    MangaErrorMessage:
      type: object
      properties:
        error:
          type: string
          enum:
            - manga_error
            - manga_error_unknown
            - manga_error_invalid_link
        trace:
          type: string
      required:
        - error
    ChapterPage:
      type: object
      properties:
        index:
          type: number
          format: float
        src:
          type: string
      required:
        - index
        - src
    ChapterPageErrorMessage:
      type: object
      properties:
        error:
          type: string
          enum:
            - chapter_error
            - chapter_error_unknown
            - chapter_error_invalid_link
            - chapter_error_fetch
            - chapter_error_no_image
        trace:
          type: string
        url:
          type: string
        index:
          type: number
          format: float
      required:
        - error
        - index
    ChapterErrorMessage:
      type: object
      properties:
        error:
          type: string
          enum:
            - chapter_error
            - chapter_error_unknown
            - chapter_error_invalid_link
            - chapter_error_fetch
            - chapter_error_no_pages
        trace:
          type: string
        url:
          type: string
      required:
        - error